modularising and commenting the whole project.  Splitting into different js modules
