+++ assets/miniGame/carTopDown.png
