+++ assets/miniGame/coin.png
