+++ styles.css
@@ -43,7 +43,10 @@ body {
   object-fit: fill;
   transform: scale(1.1);
 }
-#successMessage {
-  opacity: 0;
+#alertMessage.alert-success {
   cursor: pointer;
+}
+
+#alertMessage.alert-success:hover {
+  filter: brightness(0.95);
 }
